+++ src/components/dashboard/containers/Items/index.js
@@ -1,0 +1,1 @@
+export { default } from './Items'
